📂Class_Assignment/assignment_08/requirements.txt
streamlit==1.25.0
pandas==2.0.3
requests==2.31.0
twilio==8.5.0
tensorflow==2.12.0
pillow==9.5.0
speechrecognition==3.10.0
google-generativeai==0.8.3
python-dotenv==1.0.1
numpy>=1.24.3
plotly>=5.15.0